backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
duckdb==1.1.0
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.1
